requirements.txt
Django==3.0.3
djangorestframework==3.11.0
requests>=2.20.0
django-redis==4.3.0
redis==2.10.6